fix: store correct git state
This was somewhat confusing.

The functions in this code return 0 on success and non-zero on failure
or error. This is because there is often only one way to succeed and
many ways to fail (or error), and it's useful to have many digits to
choose from to be able to distinguish what kind of failure (or error)
the function encountered.

However, another general convention is that zero means false (failure)
and one means true (success). The struct CurrentState follows this
convention.

Somewhere along the way, I got confused and could not for the life of me
figure out why I kept on printing a non-git prompt when standing in a
git directory, and vice versa.

This commit fixes this.