feat: escape sequence validation
In terminal prompt strings like PS1 (and GP2 prompt vars), we can have
terminal escape sequences. Here are the rules they follow:
- All style manipulations start with '\033' and end with 'm'.
- To make the terminal ignore the length of these sequences, they should
  be wrapped in '\[' and '\]'.

If this isn't followed, the prompt will look strange.

This function checks the string to ensure that it doesn't, and if it
fails, it will print some informational text and return ERROR.