fix: parse and replace non-cwd tokens first
The user can configure the config parameter `string_active` and
`string_inactive` so that the text replacing the widget token can be a
new widget token.

The code supports nested tokens up to one level.

Thing is, when it blindly re-parsed the string, it returned after
replacing CWD - even if there might have been other widget tokens
afterwards.

This fixes that problem.