feat: add two new system-widgets..
.. username and hostname.

These can be used now in the prompt as dedicated widgets, so that string
length calculation can be done inside the prompt2 program. If we default
to bash's \u and \h, these take two (or zero, if I strip them) chars
each, but get expanded after the program exits - which makes calculating
the right-aligned text tricky.